% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions_1.4.R
\name{addMappa}
\alias{addMappa}
\title{Calculate mappability and append to dataframe}
\usage{
addMappa(df, binsize, GC.mappa.grange)
}
\arguments{
\item{df}{dataframe containing a feature lengths column}

\item{binsize}{The size of each bin}

\item{GC.mappa.grange}{mappability reference file}
}
\value{
dataframe containing mappability.
}
\description{
This function calculate mappability based on the bin size and appends it to
an existing dataframe.
}
