% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions_1.4.R
\name{gcmappaCorrection}
\alias{gcmappaCorrection}
\title{Correct the read counts according to the GC contents and mappability of
each bin.}
\usage{
gcmappaCorrection(df, data)
}
\arguments{
\item{df}{dataframe to contain the bin information}

\item{data}{A matrix containing the read counts. Row names are features, column names are cells}
}
\value{
A matrix containing the corrected read counts. Row names are features, column names are cells.
}
\description{
Correct the read counts according to the GC contents and mappability of
each bin.
}
