% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions_1.4.R
\name{getBinMatrix}
\alias{getBinMatrix}
\title{Calculate read counts in each bin}
\usage{
getBinMatrix(cells, df, fragments, binsize, tiles)
}
\arguments{
\item{cells}{.tsv file of cells to include in the calculation}

\item{df}{dataframe containing GCcontent, peak content and width information}

\item{fragments}{path to a .tsv.gz fragments file with corresponding .tbi file}

\item{binsize}{the size of the bin}

\item{tiles}{GRanges tiles object}
}
\value{
A matrix containing the read counts. Row names are features, column names are cells
}
\description{
This function calculates the read count of specific cells in a fragments file.
These counts are returned as a matrix
}
