% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions_1.4.R
\name{plotCNV}
\alias{plotCNV}
\title{Plot CNV heatmap}
\usage{
plotCNV(
  fc,
  data,
  ginicutoff = 0.75,
  cov_cutoff_bc,
  SNN_results,
  GMMresults,
  ref.genome = "hg38",
  k,
  project,
  a = 1,
  b = 2
)
}
\arguments{
\item{fc}{Fold change matrix}

\item{data}{Read counts before correction}

\item{ginicutoff}{Used as a cutoff for the gini index}

\item{cov_cutoff_bc}{An array of cells}

\item{SNN_results}{a dataframe with cell names and groups}

\item{GMMresults}{a dataframe with cell names, neutral proportion, neutral probability and p-value}

\item{ref.genome}{The genome reference, 'hg19' or 'hg38'}

\item{k}{Used in the plot file name to show the number of groups}

\item{project}{Used in the plot file name}

\item{a}{change the cutoff of the color bar}

\item{b}{change the cutoff of the color bar}

\item{gz}{chromosome length data, produced by seqlengths()}

\item{hg19.genes}{GRanger object that shows the genes}
}
\value{
creates heatmap and returns corresponding dataframe
}
\description{
This function creates a heatmap using CNV data, where each row is a cell
and reach column is a genomic feature. Cells can be clustered together
hierarchically based on the tree clustering result
}
