% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions_1.4.R
\name{plotCoverageDistributon}
\alias{plotCoverageDistributon}
\title{Plot Off-target Reads Coverage per cell}
\usage{
plotCoverageDistributon(data, project, cov_cutoff)
}
\arguments{
\item{data}{A matrix containing the read counts. Row names are features, column names are cells}

\item{project}{name for the plot}
}
\value{
Creates a distribution plot and return an array of cells
}
\description{
This function creates a distribution plot for off-target Reads Coverage per cell
}
