% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions_1.4.R
\name{plotReadsDistributon}
\alias{plotReadsDistributon}
\title{Plot reads per cell per bin distributons}
\usage{
plotReadsDistributon(data, project)
}
\arguments{
\item{data}{A matrix containing the read counts. Row names are features, column names are cells}

\item{project}{name for the plot}
}
\value{
Creates a distribution plot.
}
\description{
This function creates a distribution plot for reads per cell
}
