% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions_1.4.R
\name{snnTreeCluster}
\alias{snnTreeCluster}
\title{Clustering cells}
\usage{
snnTreeCluster(
  fc,
  cov_cutoff_bc,
  k = 10,
  d = 50,
  cluster_cell_cut_off,
  project
)
}
\arguments{
\item{fc}{Fold change matrix}

\item{cov_cutoff_bc}{An array of cells}

\item{k}{Used for tree clustering. Smaller 'k' usually yields finer clusters.}

\item{d}{An integer scalar specifying the number of dimensions to use for a PCA on the expression matrix prior to the nearest neighbor search.}

\item{cluster_cell_cut_off}{Remove clusters that have number of cells less than this parameter}

\item{project}{names for the file to be saved}
}
\value{
creates a dataframe with cell names and groups
}
\description{
This function clusters cells using Louvain tree cluster.
}
