% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_functions_1.4.R
\name{testfunc}
\alias{testfunc}
\title{Create offtarget and ontarget reads}
\usage{
testfunc(x)
}
\arguments{
\item{x}{number}
}
\value{
the x^2
}
\description{
This function uses the terminal to find overlap between fragments file and
peaks. Bedtools and Tabix need to be installed in order for this
preprocessing to occur
}
